Added commandline argument '--bg' to set the background color and also fill unused screen areas.
The required parameter contains 3 float numbers between 0.0 and 1.0 for each the red, green and blue component of the desired background color.
These values shall be separated by either space or by comma, similar to the syntax of the existing '--win' option.

This is an extra feature meant to add iflexibility to the similar existing '-b' option which basically now has become a synonym for '--bg 0,0,0'.

Usage:
omxplayer --bg 0,1,0.75 <filename>
- or -
omxplayer --bg "0 1 0.75" <filename>