use "type -p" instead of "which"
On Raspberry Pi's without a desktop, the following non-fatal error messages appear when starting omxplayer:

```
/usr/bin/omxplayer: line 25: which: command not found
/usr/bin/omxplayer: line 26: which: command not found
/usr/bin/omxplayer: line 27: which: command not found
```

These errors are a result of the `which` command failing to find the fbset, xset, and xrefresh binaries, which are only present when running xserver.

One way to avoid seeing the error message is to use `type -p` rather than `which`. Alternatively one could just pipe stderror to /dev/null. I chose the former, but either way works fine.