Add static title display option
The static text displayed can be set with the --title command line argument.
The default font face will be used by default with a slightly smaller font size,
these can be overridden with the --title-font and --title-font-size command line
arguments. The alignment of the title can be changed with --title-align