Forward SIGTERM and SIGINT in wrapper script
This makes sure that if the wrapper bash script receives a SIGTERM or
SIGINT, that is forwarded to the actual player to terminate it. Then,
the wrapper script does not terminate itself until the player does,
allowing cleanup to happen as normal.

This is helpful when running omxplayer from a script that wants to
terminate the player. It can just terminate the (wrapper) process that
it spawned, which will terminate the script and the player, while still
doing proper cleanup.

Note that SIGKILL cannot be handled in this way, since that kills the
script instantly with no way to trap it.